requirements.txt
tensorflow
keras
numpy
pandas
matplotlib
scikit-learn
streamlit
opencv-python
pillow
